requirements.txt
discord.py
python-dotenv
redis
aiohttp
